+++ files/default/scripts/waitfor.py
@@ -10,7 +10,8 @@ logging.basicConfig(
     level=logging.INFO
 )
 
-regex = re.compile(sys.argv[1])
+pat = sys.argv[1]
+regex = re.compile(pat)
 cmd = sys.argv[2:]
 
 
@@ -18,10 +19,15 @@ while True:
     log.info("Executing {cmd}".format(cmd=cmd))
     proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     (output, error) = proc.communicate()
+    log.info(
+        "{cmd} output: \n {output}".format(cmd=cmd, output=output)
+    )
     if regex.search(output):
         log.info("Pattern found in output")
         sys.exit(0)
     else:
-        log.info("Pattern not found, sleeping for 1s and trying again")
+        log.info(
+            "Pattern '{pat}' not found, sleeping for 1s and trying again".format(pat=pat)
+        )
         time.sleep(1)