+++ recipes/clush.rb
@@ -7,6 +7,30 @@
 
 log "\n=========== Start MapR clush.rb =============\n"
 
+# Install clush
+remote_file "Clush Package" do
+  source node['mapr']['clush']['url']
+  owner 'root'
+  group 'root'
+  path '/tmp/clustershell.tar.gz'
+end
+  
+
+
+execute "Installing clush" do
+  command "easy_install /tmp/clustershell.tar.gz"
+end
+
+remote_directory "/etc/clustershell" do
+  source "clustershell"
+  user "root"
+  group "root"
+  mode "755"
+  files_owner "root"
+  files_group "root"
+  files_mode "644"
+end
+
 all = node['mapr']['cluster_node_ips'].reject(&:empty?).join(',')
 cldb_nodes = node['mapr']['cldb_ips'].reject(&:empty?).join(',')
 zk_nodes = node['mapr']['zk_ips'].reject(&:empty?).join(',')
@@ -14,12 +38,6 @@ rm_nodes = node['mapr']['rm_ips'].reject(&:empty?).join(',')
 ws_nodes = node['mapr']['ws_ips'].reject(&:empty?).join(',')
 hs_server = node['mapr']['hs_ips']
 
-# Install clush
-ruby_block 'Installing clush' do
-  block do
-    Mixlib::ShellOut.new('rpm -ivh https://github.com/downloads/cea-hpc/clustershell/clustershell-1.6-1.el6.noarch.rpm')
-  end
-end
 
 # creating directory if it does not exist