+++ CHANGELOG.md
@@ -3,6 +3,15 @@ mapr_installation CHANGELOG
 
 This file is used to list changes made in each version of the mapr_installation cookbook.
 
+
+0.2.11
+-----
+- [eric moritz] - Added better zookeeper and cldb waiting
+
+0.2.10
+-----
+- [eric moritz] - Automated disksetup
+
 0.2.9
 -----
 - [daniel washko] - Fixing export of java_home to /etc/profile